// src/api/openapi/models/credential.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Credential { 
    id: number;
    name: string;
    password: string;
    url: string;
    username: string;
}